+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v0.1.4 (2023-12-12)
 * Added JsonApi action type
 * Improved OPTIONS request handling
 * Added matching route to Request attributes